Changed new contract permissions condition
In Redmine 2.5 the contract page was displaying without the link to "New Contract". The page was accessible by changing the URL, but it's quite a cumbersome way to add new projects. Even though the current user was set to have permissions to add contracts, the button was not displaying.

I've changed the if statement that controls who should see the add new contract button. It now uses link_to_if_authorized, which is uses in the knowledge base plugin. It displays "New Contract" button correctly on our install.